requirements.txt
cairocffi==0.7.2
CairoSVG==1.0.20
cffi==1.5.2
cssselect==0.9.1
dj-database-url==0.4.0
Django==1.9.2
django-easy-pdf==0.1.0
gunicorn==19.4.5
html5lib==0.9999999
lxml==3.5.0
Pillow==3.1.1
psycopg2==2.6.1
pycparser==2.14
PyPDF2==1.25.1
Pyphen==0.9.4
reportlab==2.7
six==1.10.0
tinycss==0.3
WeasyPrint==0.26
whitenoise==2.0.6
xhtml2pdf==0.0.6
